% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{cache_file}
\alias{cache_file}
\title{Load file from data directory.}
\usage{
cache_file(filename, prefix, sep = "_", ext = ".rds")
}
\arguments{
\item{prefix}{Analysis prefix for data}

\item{sep}{Separator (prefix sep filename)}

\item{name}{The data to save.}
}
\value{
return
}
\description{
Load file from data directory.
}

