% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{data_file2}
\alias{data_file2}
\title{Return the absolue path to a file from data directory.}
\usage{
data_file2(filename, prefix, sep = "_", ext = ".txt")
}
\arguments{
\item{filename}{The filename}
}
\value{
return
}
\description{
Return the absolue path to a file from data directory.
}

