% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-sim-features.R
\name{make_sim_features}
\alias{make_sim_features}
\title{Generate a set of features on two sets of simulations}
\usage{
make_sim_features(n, len, features, seed = 1)
}
\arguments{
\item{n}{Number of functions for each group}

\item{len}{Length of simulations}

\item{features}{Features to compute}

\item{seed}{Random generator seed.}
}
\value{
A dataframe of features, id and function names
}
\description{
Simulations are generated by sim_suite() and the 
features in the features argument are computed.a
A dataframe of features, feature name and id are returned.
}

