% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ecomplex}
\alias{plot.ecomplex}
\title{Plot the linear fit corresponding a set of   
 ecomplexity coefficients.}
\usage{
\method{plot}{ecomplex}(fit)
}
\arguments{
\item{fit}{An 'ecomplex' fit}
}
\description{
Plot the linear fit corresponding a set of   
 ecomplexity coefficients.
}

