% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_error_facet}
\alias{plot_error_facet}
\title{Facet ploto of multiple ecomplex fits}
\usage{
plot_error_facet(fits, title)
}
\arguments{
\item{fits}{Dataframe with values 'x', 'y' and labels 'id'}

\item{title}{Title of plot}
}
\description{
Facet ploto of multiple ecomplex fits
}

