% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_regression}
\alias{plot_regression}
\title{Regression Plot}
\usage{
plot_regression(df, xlab, ylab, title, confint = TRUE,
  palette = cbPalette()[2])
}
\arguments{
\item{df}{data.frame that includes columns x, y}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{palette}{A color palette}

\item{conifint}{Include confidience intervals expression(paste(alpha, " leq ", beta))}
}
\description{
Regression Plot
}

