% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot a time series or grid of time series}
\usage{
plot_ts(df, facet = TRUE, xlab = "", ylab = "", title = "", ncol = 2)
}
\arguments{
\item{df}{A data frame with columns x, y}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{ncol}{grid columns}
}
\description{
Plot a time series or grid of time series
}

