% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-rf-classify-segments.R
\name{predict_single_ch}
\alias{predict_single_ch}
\title{Leave one out classification of a single trial}
\usage{
predict_single_ch(df_, data_, tnum, train, colnums, thresh = 0.45)
}
\arguments{
\item{data_}{: Segment book-keeping data}

\item{df:}{Features means for a given trial}

\item{tindex}{: The trial index (refers to the filtered index)}

\item{colnums:}{df_means columns to use in classification}
}
\value{
A list of 1. Prediction probability vector, 
                  2. Prediction vector
                  3. Weight vector
}
\description{
Takes the mean of features on segmented data and
 predicts segments using random forest classifier
}

