% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{save_cache}
\alias{save_cache}
\title{Save processed data to a directory.}
\usage{
save_cache(dat, filename, prefix, sep = "_", ext = ".rds")
}
\arguments{
\item{filename}{The local filename.}

\item{data}{The data to save.}
}
\value{
return
}
\description{
Save processed data to a directory.
}

