% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{save_plot}
\alias{save_plot}
\title{Save most recent image using ggsave}
\usage{
save_plot(filename, prefix, sep = "_", ext = ".pdf")
}
\arguments{
\item{filename}{The local file name.}
}
\value{
return
}
\description{
Save most recent image using ggsave
}

