% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatterplot_features}
\alias{scatterplot_features}
\title{Scatter plot of features}
\usage{
scatterplot_features(xname, yname, df_, id = NULL)
}
\arguments{
\item{xname}{Column name for x variable}

\item{yname}{Column name for y variable}

\item{df_}{A data frame with xname, yname columns}

\item{ids}{}
}
\description{
Scatter plot of features
}

