% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stemplot}
\alias{stemplot}
\title{Create Matlab-like stemplot}
\usage{
stemplot(x, y, pch = 16, linecol = 1, ...)
}
\arguments{
\item{x}{A numeric vector in index points}

\item{y}{A numeric vector of function points}

\item{pch}{Leaf symbol}

\item{linecol}{}

\item{clinecol}{}
}
\description{
Create Matlab-like stemplot
}

